Add command Id normalizer support
IdNormalizer now returns an int (as a string) mimicking the base class

Fix bug where an exception was thrown instead of returning null when allowNull was true

Break normalizer tests out into classes per type
